Sys numerics now uses doubles instead of floats (#101)
***NO_CI***

+++ src/System.Numerics/Vector2_Intrinsics.cs
@@ -15,11 +15,11 @@ namespace System.Numerics
         /// <summary>
         /// The X component of the vector.
         /// </summary>
-        public Single X;
+        public double X;
         /// <summary>
         /// The Y component of the vector.
         /// </summary>
-        public Single Y;
+        public double Y;
 
         #region Constructors
         /// <summary>
@@ -27,7 +27,7 @@ namespace System.Numerics
         /// </summary>
         /// <param name="value">The element to fill the vector with.</param>
         
-        public Vector2(Single value) : this(value, value) { }
+        public Vector2(double value) : this(value, value) { }
 
         /// <summary>
         /// Constructs a vector with the given individual elements.
@@ -35,7 +35,7 @@ namespace System.Numerics
         /// <param name="x">The X component.</param>
         /// <param name="y">The Y component.</param>
         
-        public Vector2(Single x, Single y)
+        public Vector2(double x, double y)
         {
             X = x;
             Y = y;
@@ -47,7 +47,7 @@ namespace System.Numerics
         /// Copies the contents of the vector into the given array.
         /// </summary>
         /// <param name="array">The destination array.</param>
-        public void CopyTo(Single[] array)
+        public void CopyTo(double[] array)
         {
             CopyTo(array, 0);
         }
@@ -60,7 +60,7 @@ namespace System.Numerics
         /// <exception cref="ArgumentOutOfRangeException">If index is greater than end of the array or index is less than zero.</exception>
         /// <exception cref="ArgumentException">If number of elements in source vector is greater than those available in destination array
         /// or if there are not enough elements to copy.</exception>
-        public void CopyTo(Single[] array, int index)
+        public void CopyTo(double[] array, int index)
         {
             if (array == null)
             {
@@ -99,7 +99,7 @@ namespace System.Numerics
         /// <param name="value2">The second vector.</param>
         /// <returns>The dot product.</returns>
         
-        public static float Dot(Vector2 value1, Vector2 value2)
+        public static double Dot(Vector2 value1, Vector2 value2)
         {
             return value1.X * value2.X +
                    value1.Y * value2.Y;
@@ -152,7 +152,7 @@ namespace System.Numerics
         
         public static Vector2 SquareRoot(Vector2 value)
         {
-            return new Vector2((Single)Math.Sqrt(value.X), (Single)Math.Sqrt(value.Y));
+            return new Vector2(Math.Sqrt(value.X), Math.Sqrt(value.Y));
         }
         #endregion Public Static Methods
 
@@ -200,7 +200,7 @@ namespace System.Numerics
         /// <param name="right">The source vector.</param>
         /// <returns>The scaled vector.</returns>
         
-        public static Vector2 operator *(Single left, Vector2 right)
+        public static Vector2 operator *(double left, Vector2 right)
         {
             return new Vector2(left, left) * right;
         }
@@ -212,7 +212,7 @@ namespace System.Numerics
         /// <param name="right">The scalar value.</param>
         /// <returns>The scaled vector.</returns>
         
-        public static Vector2 operator *(Vector2 left, Single right)
+        public static Vector2 operator *(Vector2 left, double right)
         {
             return left * new Vector2(right, right);
         }
@@ -236,9 +236,9 @@ namespace System.Numerics
         /// <param name="value2">The scalar value.</param>
         /// <returns>The result of the division.</returns>
         
-        public static Vector2 operator /(Vector2 value1, float value2)
+        public static Vector2 operator /(Vector2 value1, double value2)
         {
-            float invDiv = 1.0f / value2;
+            double invDiv = 1.0 / value2;
             return new Vector2(
                 value1.X * invDiv,
                 value1.Y * invDiv);
